requirements.txt
altgraph==0.17.2
appnope==0.1.3
appscript==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
confection==0.0.1
cryptography==36.0.1
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docxcompose==1.3.4
docxtpl==0.15.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.0/en_core_web_md-3.4.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fastjsonschema==2.16.2
feedparser==6.0.8
fonttools==4.29.1
funcy==1.17
future==0.18.2
gensim==4.2.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
kiwisolver==1.3.2
langcodes==3.3.0
lxml==4.7.1
macholib==1.16
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.4
murmurhash==1.0.8
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numexpr==2.8.3
numpy==1.22.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pefile==2021.9.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
preshed==3.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pyLDAvis==3.3.1
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySimpleGUI==4.56.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.2.1
regex==2022.9.13
requests==2.27.1
retrying==1.3.3
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.30
sqlalchemy-views==0.3.1
srsly==2.4.4
stack-data==0.5.0
terminado==0.16.0
thinc==8.1.2
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.19.0
trio-websocket==0.9.2
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.8
wasabi==0.10.1
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
wordcloud==1.8.2.2
wsproto==1.0.0
xlwings==0.27.7
